requirements.txt
streamlit
numpy
opencv-python
tensorflow
Pillow
scikit-image
keras
gdown